init global UseBiasedLocking with product dependent value false on aarch64
n.b. this deliberately breaks non-aarch64 builds (because they do not
provide the product_pd declaration for this property). the intention
of this patch is merely to switch this setting off temporarily while
bootstrapping locking on aarch64. leaving other builds broken should
ensure we back out this change before any attempt to integrate is
made.