fixed 4 problems with signature handler code
the pass_xxx routines default case handlers were not incrementing the
passed gp and fp counts, num_int_args and num_fp_args. this meant that
the call format passed to the brx86 call was being computed
incorrectly never reporting more than 8 gp args.

the pass_object routine default case handler (more than 6 int args)
was testing the wrong value (it checked temp which was set to 0 by the
preceding instruction) and hence always skipping the store. fixing it
to use the right value (the contents of the address in r0) is still
not enough because the code needs to always do the store with either
the value in r0 or 0.

the initial stack offset was set to 8 as per intel (to avoid
clobbering a return address) but actually it should be 0

stacked args were being stored relative to the Java stack pointer but
the caller expects to see them relative to the machine sp.

with these fixes the call to the garphics routine which takes 11
arguments works correctly.

n.b. this commit is just an interim patch while I work on the next bug
this uncovers. this code really needs patching to use the arm
conditional moves.