support for compiled code breaks
added callouts to notify simulator of generation and subsequent relocation of compiled C1 code

modified bccheck callback so it uses a pc and fp to establish the
desired method name. specifically, fp is used when the pc is in
interpreter code to establish whether to use registers rmethod and
rbcp (fp == 0) or the stack saved values (fp != 0).