fixed oopmap at C2 safepoint read with dependent change to c1 safepoint check
Safepoint in C2 cannot be generated as a macro sequence "load the poll
address; read from it" because the generic C2 code can only place the
oop map at the start of such a sequence. thsi is no good because the
SEGV trap handler relies on the oop map being placed at the read.

this means we have to configure the C2 rule for a Safepoint node to
accept the node address as an input and let C2 plant the address load
(constant load) as a prior step. this means the address load will not
always directly precede the read.

this doe snot change c1 per se. it can still plant these instructions
adjacent to each other. however, the native instruction safepoint
check and the reloc code need to be modified:

i) to accept a reloc_a pollah on a ldr(zr, [reg, off]) instruction but
do no relocation on it rather than look for a preceding reloc address
load to relocate

ii) to relocate the constant load (adrp or mov64) merely by virtue of
it being marked as reloc_poll